Backport preferred_stride control from newer JetPack
Use preferred_stride control to set the bytesperline value to be 64B
align, to support the resolutions those width is not 64B align. This
requirement is not seen in documents but mentioned in NVIDIA's forum.
VI driver handles this and the v4l2 format's bytesperline is set too.

Modify the GUI app streamApp to set preferred_stride automatically for
the stream width.

Signed-off-by: Xin Zhang <xin.x.zhang@intel.com>