Fixed narrowing conversion of imu framerates
The variable type on the definition of all framerates is u16.
But in ds5_mux_s_frame_interval() and __ds5_probe_framerate(),
the framerate will be narrowed to u8. Especially for IMU, its
framerate likely exceeds the maximum value of u8.

Fixed this by changing u8 to u16.
Another change is adding 50 to imu framerates array.

Signed-off-by: Shikun Ding <shikun.ding@intel.com>